Added the Possibility to Skip Subtitle Download

+++ src/lib/utils.js
@@ -13,19 +13,23 @@ var utils = {
 	
 	eventMod: require('events'),
 	
-	download: function(url, dest, cb) {
-	  var http = require('http');
-	  var fs = require('fs');
-	  var file = fs.createWriteStream(dest);
-	  var request = http.get(url, function(response) {
-		response.pipe(file);
-		file.on('finish', function() {
-		  file.close(cb);  // close() is async, call cb after close completes.
-		});
-	  }).on('error', function(err) { // Handle errors
-		fs.unlink(dest); // Delete the file async. (But we don't check the result)
-		if (cb) cb(err.message);
-	  });
+	download: function(url, dest, cb, skip) {
+		if (skip) {
+			cb('skipped');
+		} else {
+			var http = require('http');
+			var fs = require('fs');
+			var file = fs.createWriteStream(dest);
+			var request = http.get(url, function(response) {
+				response.pipe(file);
+				file.on('finish', function() {
+					file.close(cb);  // close() is async, call cb after close completes.
+				});
+			}).on('error', function(err) { // Handle errors
+				fs.unlink(dest); // Delete the file async. (But we don't check the result)
+				if (cb) cb(err.message);
+			});
+		}
 	},
 
 	validateUrl: function(torLink) {
